// src/store/auth/auth.types.ts
export enum AUTH_ACTION_TYPES {
  GOOGLE_SIGN_IN = 'auth/GOOGLE_SIGN_IN',
  EMAIL_AND_PASSWORD_SIGN_IN = 'auth/EMAIL_AND_PASSWORD_SIGN_IN',
  EMAIL_AND_PASSWORD_SIGN_UP = 'auth/EMAIL_AND_PASSWORD_SIGN_UP',
  SIGN_IN_SUCCESS = 'auth/SIGN_IN_SUCCESS',
  FAILURE = 'auth/FAILURE',
  CHECK_USER_SESSION = 'auth/CHECK_USER_SESSION',
  SIGN_OUT = 'auth/SIGN_OUT',
  SIGN_OUT_SUCCESS = 'auth/SIGN_OUT_SUCCESS',
}